rename library to resource
