+++ app/src/main/AndroidManifest.xml
@@ -76,7 +76,7 @@
             android:name=".ui.news.ReplyActivity"
             android:theme="@style/AppTheme" />
         <activity
-            android:name=".ui.library.AddResourceActivity"
+            android:name=".ui.resources.AddResourceActivity"
             android:theme="@style/AppTheme" />
         <activity android:name=".ui.viewer.AudioPlayerActivity" />
         <activity
