+++ app/src/main/java/org/ole/planet/myplanet/ui/course/CourseFragment.kt
@@ -26,7 +26,7 @@ import org.ole.planet.myplanet.model.RealmRating.Companion.getRatings
 import org.ole.planet.myplanet.model.RealmSearchActivity
 import org.ole.planet.myplanet.model.RealmTag
 import org.ole.planet.myplanet.model.RealmTag.Companion.getTagsArray
-import org.ole.planet.myplanet.ui.library.CollectionsFragment
+import org.ole.planet.myplanet.ui.resources.CollectionsFragment
 import org.ole.planet.myplanet.utilities.KeyboardUtils.setupUI
 import org.ole.planet.myplanet.utilities.Utilities
 import java.util.Calendar
