+++ app/src/main/java/org/ole/planet/myplanet/ui/resources/AdapterLibrary.kt
@@ -1,4 +1,4 @@
-package org.ole.planet.myplanet.ui.library
+package org.ole.planet.myplanet.ui.resources
 
 import android.content.Context
 import android.text.TextUtils
