+++ app/src/main/java/org/ole/planet/myplanet/ui/resources/LibraryDetailFragment.kt
@@ -1,4 +1,4 @@
-package org.ole.planet.myplanet.ui.library
+package org.ole.planet.myplanet.ui.resources
 
 import android.os.Bundle
 import android.text.TextUtils
