+++ app/src/main/java/org/ole/planet/myplanet/ui/resources/LibraryFragment.kt
@@ -1,4 +1,4 @@
-package org.ole.planet.myplanet.ui.library
+package org.ole.planet.myplanet.ui.resources
 
 import android.app.AlertDialog
 import android.content.DialogInterface
