+++ app/src/main/res/layout/fragment_add_resource.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.library.AddResourceFragment">
+    tools:context=".ui.resources.AddResourceFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
