+++ app/src/main/res/layout/fragment_library_detail.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.library.LibraryDetailFragment">
+    tools:context=".ui.resources.LibraryDetailFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
