+++ app/src/main/res/layout/fragment_library_filter.xml
@@ -7,7 +7,7 @@
     android:background="@color/md_white_1000"
     android:orientation="vertical"
     android:padding="@dimen/padding_normal"
-    tools:context=".ui.library.LibraryFilterFragment">
+    tools:context=".ui.resources.LibraryFilterFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
