+++ app/src/main/res/layout/fragment_my_course.xml
@@ -5,7 +5,7 @@
     android:id="@+id/my_course_parent_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.library.LibraryFragment">
+    tools:context=".ui.resources.LibraryFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
